+++ source/GameInterface.h
@@ -6,7 +6,7 @@ class GameInterface
 {
 public:
 	virtual void Initialize(RenderComponentInterface* p_renderComponentInterface) = 0;
-	virtual void Update(double p_deltaTime, float p_mousePositionX, float p_mousePositionY /* add keyboard parameters here*/) = 0;	// TODO add vector with keyclicks here
+	virtual void Update(double p_deltaTime, int p_mousePositionX, int p_mousePositionY /* add keyboard parameters here*/) = 0;	// TODO add vector with keyclicks here
 	virtual void Render() = 0;
 };
